tested

+++ libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: yzhang2 <yzhang2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 18:32:25 by yzhang2           #+#    #+#             */
-/*   Updated: 2025/04/28 18:42:20 by yzhang2          ###   ########.fr       */
+/*   Updated: 2025/05/24 16:30:02 by yzhang2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *s1)
 {
 	char	*str;
-	int		length;
+	size_t	length;
 
 	length = ft_strlen(s1) + 1;
 	str = malloc(sizeof(char) * length);
